requirements.txt
# Core Dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Deep Learning (Optional)
torch>=1.9.0
transformers>=4.11.0

# API and Web
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.26.0

# Data Processing
python-dotenv>=0.19.0
pyyaml>=5.4.1

# Testing
pytest>=6.2.5
pytest-cov>=2.12.0

# Logging and Monitoring
loguru>=0.5.3

# Development Tools
black>=21.7b0
isort>=5.9.3
flake8>=3.9.2
